backend/engagement_classifier/model.py
"""
Model placeholder for future deep learning engagement classifier
This file will contain the actual model loading logic when ML is integrated
"""
from typing import Any, Optional
import numpy as np


class EngagementClassifierModel:
    """
    Placeholder class for future engagement classification model
    
    In production, this will load a trained deep learning model
    that takes student data as input and returns engagement level
    """
    
    def __init__(self, model_path: Optional[str] = None):
        """
        Initialize model (currently placeholder)
        
        Args:
            model_path: Path to trained model file
        """
        self.model = None
        self.model_path = model_path
        self.is_loaded = False
        
        if model_path:
            self.load_model()
    
    def load_model(self):
        """
        Load trained engagement classification model
        
        Currently a placeholder - will be implemented with actual ML model
        """
        # TODO: Implement actual model loading
        # Example:
        # import tensorflow as tf
        # self.model = tf.keras.models.load_model(self.model_path)
        # self.is_loaded = True
        
        print("WARNING: Using placeholder model. Actual ML model not yet loaded.")
        self.is_loaded = False
    
    def predict(self, features: np.ndarray) -> str:
        """
        Predict engagement level from features
        
        Args:
            features: Array of student features (response_time, is_correct, etc.)
        
        Returns:
            Predicted engagement level: "active", "moderate", or "passive"
        """
        if not self.is_loaded or self.model is None:
            raise RuntimeError("Model not loaded. Use rules-based classifier.")
        
        # TODO: Implement actual prediction
        # prediction = self.model.predict(features)
        # engagement_level = interpret_prediction(prediction)
        # return engagement_level
        
        pass
    
    def preprocess(self, raw_data: dict) -> np.ndarray:
        """
        Preprocess raw student data for model input
        
        Args:
            raw_data: Dictionary with student metrics
        
        Returns:
            Preprocessed numpy array
        """
        # TODO: Implement data preprocessing
        # features = extract_features(raw_data)
        # normalized = normalize(features)
        # return normalized
        
        pass

