backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
websockets==12.0
apscheduler==3.10.4
pyjwt==2.8.0
requests==2.31.0
python-dotenv==1.0.0
PyMuPDF==1.23.8
python-pptx==0.6.21
nltk==3.8.1
pillow==10.1.0
alembic==1.12.1

